pyproject.toml
[project]
name = "venus-os-fronius-proxy"
version = "0.1.0"
description = "Modbus TCP proxy making SolarEdge inverters appear as Fronius to Venus OS"
requires-python = ">=3.11"
dependencies = [
    "pymodbus>=3.6,<4.0",
    "structlog>=24.0",
    "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
